% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-ggplot.R
\name{cumsum_na}
\alias{cumsum_na}
\title{Cumsum vector with NA values.}
\usage{
cumsum_na(x, ...)
}
\arguments{
\item{x}{The vector for which the cumulative sum is desired.}

\item{...}{Additional arguments for the \code{cumsum} function.}
}
\description{
\code{cumsum_na} Cumulative sum for vectors with NA values.
}
\details{
When missing values are present, \code{cumsum_na} replace missing values for 0 and
compute the usual \code{cumsum}.
}
\author{
Erick A. Chacon-Montalvan
}
