% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-extremes.R
\name{find_flood_drought}
\alias{find_flood_drought}
\title{Identify floods and droughts based on SPI.}
\usage{
find_flood_drought(spi)
}
\arguments{
\item{spi}{A vector series of standardized precipitation values.}
}
\value{
return.
}
\description{
\code{find_flood_drought} identifies floods and droughts for a vector series of
standardized precipitation values.
}
\details{
Floods (droughts) are identified by a subset of negative values where at least one value
exceeds -1 (1).
}
\author{
Erick A. Chacon-Montalvan
}
