% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi.R
\name{ggplot.mbsi}
\alias{ggplot.mbsi}
\title{Plot fitting}
\usage{
\method{ggplot}{mbsi}(data)
}
\arguments{
\item{par.}{}
}
\value{
return.
}
\description{
\code{function} description.
}
\details{
details.
}
\examples{


}
\author{
Erick A. Chacon-Montalvan
}
