% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbsi-package.R
\docType{package}
\name{mbsi-package}
\alias{mbsi-package}
\alias{_PACKAGE}
\title{mbsi: model-based standardized index}
\description{
The \code{mbsi} package provide tools to compute and visualize extreme
hydro-climatic events using the standardized precipitation index (SPI) and the
model-based standardized index (MBSI). The difference with between the MBSPI and
the classical SPI index is that it consider the association between continuous
times using cycle P-splines \code{pbc}. The package can also with work with
precipitation series containing missing values (NA), 0 or only non-zero values.
}
\author{
Erick A. Chacon-Montalvan
}
