% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-mbsi.R
\name{mbsi}
\alias{mbsi}
\title{Compute the model-based standardized index (MBSI)}
\usage{
mbsi(y, time, tscale = 1, period = 365/7)
}
\arguments{
\item{y}{Precipitation level or other variable under study.}

\item{time}{Time associated with the variable \code{y}. This variable can be in
weeks, months, etc.}

\item{tscale}{Time-scale to compute the MBSI in \code{time} units. This argument
is used to return the standardized values at this scale. Usually defined to
identify different types of droughts of floods.}

\item{period}{Numeric value representing the period to define seasonality. For
example, 53 weeks, 12 months, 365 days. As it can be seen, it depends of the units
of the argument \code{time}.}
}
\value{
A dataframe consisting of \code{y}, \code{time}, \code{season}, \code{mu},
\code{sigma}, \code{pzero}, \code{ecdf} and \code{spi}.
}
\description{
\code{mbsi} computes the model-based standardized index using \code{gamlss}
models. The difference with the classical \code{spi} index is that it consider the
association between continuous times using cycle P-splines \code{pbc}.
It can also with work with precipitation series containing NA, 0 or
only non-zero values.
}
\details{
details.
}
\examples{

data(simrain)
spi_rain <- mbsi(simrain$rain, simrain$time)

# Visualize model fitting
plot(spi_rain)
# Visualize distribution of empirical cumulative density function
plot(spi_rain, which = "ecdf", binwidth = 0.05)
# Visualize extreme events
plot_extremes(spi_rain, threshold = 2)

}
\author{
Erick A. Chacon-Montalvan
}
