% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-ggplot.R
\name{plot_extremes}
\alias{plot_extremes}
\title{Plot floods and droughts episodes using standardized precipitation values}
\usage{
plot_extremes(data, threshold = 1)
}
\arguments{
\item{data}{A \code{mbsi} object returned by \code{spi} or \code{mbsi}, which
contains the standardized precipitation values.}

\item{threshold}{An numeric value used to detect the extreme events, 1 (-1) is
used by default as proposed by Mckee et al. (1993).}
}
\value{
A \code{ggplot} object. The graph is shown when this object is printed.
}
\description{
\code{plot_extremes} Plot floods and droughts episodes using the function
\code{find_flood_drought}. This function employs the
definition proposed by Mckee et al. (1993) using a time series of standardized
precipitation values. A drought (flood) is defined  as a period of time in
which the SPI is continuously negative (positive) reaching at least
one value lower (higher) or equal to -1 (1).
}
\details{
Although the definition of Mckee et al. (1993) uses a threshold of 1 (-1). Other
values can be used for this threshold.
}
\examples{

data(simrain)
spi_rain <- mbsi(simrain$rain, simrain$time)

# Visualize extreme events
plot_extremes(spi_rain, threshold = 2)
plot_extremes(spi_rain, threshold = 1.5)
plot_extremes(spi_rain, threshold = 1)

}
\author{
Erick A. Chacon-Montalvan
}
