% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simrain-data.R
\docType{data}
\name{simrain}
\alias{simrain}
\title{Simulated rainfall data}
\format{An object of class \code{data.frame} with 520 rows and 6 columns.}
\usage{
data(simrain)
}
\description{
Simulated rainfall data to be used in the examples of \code{mbsi} package. This
data has a seasonal component and a serially correlated component.
}
\examples{

data(simrain)
plot(simrain$time, simrain$rain, type = "b")
lines(simrain$time, exp(-2 + simrain$season), col = 2)

}
\author{
Erick A. Chacon-Montalvan
}
\keyword{datasets}
