% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-classic.R
\name{spi}
\alias{spi}
\title{Classic standardized precipitation index (SPI)}
\usage{
spi(y, time, tscale = 1, period = 52)
}
\arguments{
\item{y}{Usually rainfall from which to compute the SPI, but other variables like
precipitation can also be used.}

\item{time}{Time indexed to the values of the argument \code{y}, it can be in weeks,
months, etc.}

\item{tscale}{Time-scale to compute the SPI in time units. This argument is used
to perform a moving average to the rainfall series before computing the SPI.}

\item{period}{Numeric value representing the period to define seasonality. For
example, 53 weeks, 12 months, 365 days. As it can be seen, it depends of the units
of the argument \code{time}.}
}
\value{
A dataframe consisting of \code{y}, \code{time}, \code{season}, \code{mu},
\code{sigma}, \code{pzero}, \code{ecdf} and \code{spi}.
}
\description{
\code{spi} computes the spi index using \code{gamlss} models.
It can also with work with precipitation series containing NA, 0 or
only non-zero values.
}
\details{
details.
}
\examples{

data(simrain)
spi_rain <- spi(simrain$rain, simrain$time)

# Visualize model fitting
plot(spi_rain)
# Visualize distribution of empirical cumulative density function
plot(spi_rain, which = "ecdf", binwidth = 0.05)
# Visualize extreme events
plot_extremes(spi_rain, threshold = 2)

}
\author{
Erick A. Chacon-Montalvan
}
