% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb-si.R
\name{spi_week}
\alias{spi_week}
\title{Compute the Standardized Precipitation Index.}
\usage{
spi_week(rain, tscale = 1, period = 365/7, package = "gamlss",
  plot = FALSE)
}
\arguments{
\item{rain}{Precipitation level.}

\item{tscale}{Time-scale to compute the SPI in week units.}

\item{period}{Period (e.g. 53 weeks) defined to model the seasonal effect.}

\item{package}{Package to use. \code{gamlss} for backfitting and \code{bamlss} for baysian gamlss models.}

\item{plot}{Logical value indicating if a plot should be draw or not.}
}
\value{
dataframe consisting of spi, mu, sigma and pzero.
}
\description{
\code{spi_week} computes the weekly spi index with gamlss models.
It can also with work with precipitation series containing NA, 0 or
only non-zero values.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}
