% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-ggplot.R
\name{stat_events}
\alias{stat_events}
\title{Extreme events.}
\usage{
stat_events(mapping = NULL, data = NULL, geom = "ribbon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping created by \code{aes} or \code{aes}.}

\item{data}{Dataset to use.}

\item{geom}{Geometry.}

\item{position}{Position.}

\item{na.rm}{Logical value to remove missing values.}

\item{show.legend}{Show legend.}

\item{inherit.aes}{Inherit aesthetics.}

\item{...}{Additional arguments passed to \code{layer}.}
}
\description{
\code{stat_events} computes a data frame to adequatly draw areas
corresponding to extreme events in a time series.
}
\examples{

# Obtain a zero-mean time series
library(ggplot2)
economics <- transform(economics,
  unemploy_zero = 2 * (unemploy - mean(unemploy)) / sd(unemploy))

# Plot time series with appropiate shade for negative and positive values
ggplot(economics, aes(date, unemploy_zero)) +
  geom_line() +
  geom_point(color = 2, alpha = 0.3) +
  stat_events(alpha = 0.3)

# Plot time series with appropiate shade for a given threshold
ggplot(economics, aes(date, unemploy_zero)) +
  geom_line() +
  geom_point(color = 2, alpha = 0.3) +
  stat_events(threshold = 0.5, alpha = 0.3)

# Plot time series with appropiate shade for a given threshold
ggplot(economics, aes(date, unemploy_zero)) +
  geom_line() +
  stat_events(aes(event = I(1 * (unemploy_zero > 1)), fill = "positive peak"),
              threshold = 1, alpha = 0.3) +
  stat_events(aes(event = I(1 * (unemploy_zero < -1)), fill = "negative peak"),
              threshold = -1, alpha = 0.3)

}
\author{
Erick A. Chacon-Montalvan
}
