% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes-detect.R
\name{find_flood_drought}
\alias{find_flood_drought}
\title{Detect floods and droughts based on Mckee et al. (1993)}
\usage{
find_flood_drought(spi, threshold = qnorm(1 - 0.05/2), labels = c("normal",
  "drought", "flood"))
}
\arguments{
\item{spi}{A vector representing the time series of the standardized values.}

\item{threshold}{An numeric value used to detect the extreme events, 1 (-1) is
used by default as proposed by Mckee et al. (1993).}

\item{labels}{A character vector providing the labels used to return the three
types of events \code{normal}, \code{drought} and \code{flood}.}
}
\value{
A factor vector indicating the time of event the standardized
precipitation value corresponds.
}
\description{
\code{find_flood_drough} detect events of floods and droughts based on the
definition proposed by Mckee et al. (1993) using a time series of standardized
precipitation values. A drought (flood) is defined  as a period of time in
which the SPI is continuously negative (positive) reaching at least
one value lower (higher) or equal to -1 (1).
}
\details{
Although the definition of Mckee et al. (1993) uses a threshold of 1 (-1). Other
values can be used for this threshold.
}
\examples{

data(simrain)
# Compute standardized values
spi_rain <- mbsi(simrain$rain, simrain$time)
# Identify floods and droughts
spi_rain$event <- find_flood_drought(spi_rain$spi, threshold = 2)
spi_rain$event

}
\author{
Erick A. Chacon-Montalvan
}
