% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-mbsi.R
\name{mbsi}
\alias{mbsi}
\title{Compute the model-based standardized index (MBSI)}
\usage{
mbsi(y, time, tscale = 1, period = 365/7)
}
\arguments{
\item{y}{Precipitation level or other variable under study.}

\item{time}{Time associated with the variable \code{y}.}

\item{tscale}{Time-scale to compute the MBSI in \code{time} units.}

\item{period}{Period (e.g. 53 weeks) defined to model the seasonal effect.}
}
\value{
A dataframe consisting of \code{y}, \code{time}, \code{season}, \code{mu},
\code{sigma}, \code{pzero}, \code{ecdf} and \code{spi}.
}
\description{
\code{mbsi} computes the model-based standardized index using \code{gamlss}
models. The difference with the classical \code{spi} index is that it consider the
association between continuous times using cycle P-splines \code{pbc}.
It can also with work with precipitation series containing NA, 0 or
only non-zero values.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}
