% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-mbsi.R
\name{plot.mbsi}
\alias{plot.mbsi}
\title{Plot fitting when computing the SPI}
\usage{
\method{plot}{mbsi}(x, which = c("fit", "ecdf"), binwidth = 0.05, ...)
}
\arguments{
\item{data}{An \code{mbsi} object returned by \code{spi}.}
}
\value{
A \code{ggplot} object. The graph is shown when this object is printed.
}
\description{
\code{plot.mbsi} make a graph of the mean and coverage interval obtained when
computing the \code{spi}. This is useful to evaluate the seasonal behaviour and
the parameter estimation made by the classical SPI.
}
\details{
details.
}
\examples{


}
\author{
Erick A. Chacon-Montalvan
}
